requirements.txt
aiohttp==3.11.14
fastapi==0.115.5
faiss-cpu==1.10.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.6
langchain-together==0.3.0
numpy==1.26.4
openai==1.66.3
pydantic==2.10.2
requests==2.32.3
sentence-transformers==3.4.1
uvicorn>=0.22.0
python-dotenv==1.0.1
beautifulsoup4